Rename some GraphQL types related to stake
