+++ NineChronicles.Headless.Tests/GraphTypes/Abstractions/StakeRewardsTypeTest.cs
@@ -2,13 +2,14 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using GraphQL;
 using GraphQL.Execution;
+using NineChronicles.Headless.GraphTypes.Abstractions;
 using NineChronicles.Headless.GraphTypes.States.Models.Table;
 using Xunit;
 using static NineChronicles.Headless.Tests.GraphQLTestUtils;
 
-namespace NineChronicles.Headless.Tests.GraphTypes.States.Models
+namespace NineChronicles.Headless.Tests.GraphTypes.Abstractions
 {
-    public class StakeRegularRewardSheetTypeTest
+    public class StakeRewardsTypeTest
     {
         [Fact]
         public async Task Query()
@@ -28,7 +29,7 @@ namespace NineChronicles.Headless.Tests.GraphTypes.States.Models
                     }
                 }
             }";
-            var queryResult = await ExecuteQueryAsync<StakeRegularRewardSheetType>(
+            var queryResult = await ExecuteQueryAsync<StakeRewardsType>(
                 query,
                 source: (Fixtures.TableSheetsFX.StakeRegularRewardSheet,
                     Fixtures.TableSheetsFX.StakeRegularFixedRewardSheet)
